\name{as.scp}
\alias{as.scp}
\title{transfer a marked ppp object into a scp object, it mark is just the species name}
\usage{
  as.scp(data.ppp)
}
\arguments{
  \item{data.ppp}{an ppp object with species name as its
  mark}
}
\description{
  transfer a marked ppp object into a scp object, it mark
  is just the species name
}

