\name{backwardStep}
\alias{backwardStep}
\title{To automatically select a best fitted clustering model based on pvalue}
\usage{
  backwardStep(fittedmodel, siglevel = 0.05)
}
\arguments{
  \item{fittedmodel}{a fm object represent a fitted
  clustering model for a point pattern}

  \item{siglevel}{a marginal significance level used to
  decide which part of model is significance.}
}
\description{
  To automatically select a best fitted clustering model
  based on pvalue
}

