\name{changeTostar}
\alias{changeTostar}
\title{A help functions to change pvalue into the correspondence significant level using *}
\usage{
  changeTostar(x)
}
\arguments{
  \item{x}{the pvalue}
}
\description{
  A help functions to change pvalue into the correspondence
  significant level using *
}
\examples{
changeTostar(1)

changeTostar(0.06)

changeTostar(0.03)

changeTostar(0.001)
}

