\name{comdistnt_C}
\alias{comdistnt_C}
\title{An C++ implementation of inter-community mean nearest taxon distance}
\usage{
  comdistnt_C(comm, dis, abundance.weighted = FALSE,
    exclude.conspecifics = FALSE)
}
\description{
  An C++ implementation of inter-community mean nearest
  taxon distance
}

