\name{dataframeToIms}
\alias{dataframeToIms}
\title{convert a data.frame into a list of im objects}
\usage{
  dataframeToIms(data, size, xcol, yrow)
}
\arguments{
  \item{data}{a data.frame object contains observed
  variables}

  \item{size}{size of the observed variable matrix}

  \item{xcol}{ector of x coordinates for the pixel gid}

  \item{yrow}{ector of y coordinates for the pixel gid}
}
\value{
  a list of im objects
}
\description{
  convert a data.frame into a list of im objects
}

