\name{ecol_phylo_corr}
\alias{ecol_phylo_corr}
\title{generate phylogenetic relationship accroding to the ecological relationship}
\usage{
  ecol_phylo_corr(ecological_similarity, R2)
}
\arguments{
  \item{ecological_similarity}{a matrix represent the
  ecological distance between species}

  \item{R2}{the expected adjust R square of the correlation
  between ecological dissimlarity and ecological distance}
}
\description{
  generate phylogenetic relationship accroding to the
  ecological relationship
}

