\name{example_sce_f1}
\alias{example_sce_f1}
\title{this is an example of puting you R functiono into sce package}
\usage{
  example_sce_f1(x, y)
}
\arguments{
  \item{x}{a numeric vector.}

  \item{y}{a numeric vector or an interge.}
}
\description{
  this is an example of puting you R functiono into sce
  package
}
\details{
  This is the detailed introduction about this function
}
\examples{
example_sce_f1(1:4,1:4)
}

