\name{HSindex}
\alias{HSindex}
\title{The Hopkins and skellam's aggregation index}
\usage{
  HSindex(sp.ppp, nfocals, k = 1, q = 1,
    bordercorrection = 0)
}
\arguments{
  \item{sp.ppp}{an ppp object of a population}

  \item{nfocals}{the number of focal points/events}

  \item{k}{the order of neighbor}

  \item{q}{the number of equal angle sector}

  \item{bordercorrection}{the distance from the border
  without focal points}
}
\description{
  The Hopkins and skellam's aggregation index
}

