\name{pcaHabitat}
\alias{pcaHabitat}
\title{PCA transformation of the habitat variables}
\usage{
  pcaHabitat(habitat, scale = TRUE)
}
\arguments{
  \item{habitat}{a list of im with the same dim}

  \item{scale}{scale all variables under the unit
  variance.}
}
\description{
  PCA transformation of the habitat variables
}

