\name{pcf_adaptive}
\alias{pcf_adaptive}
\title{estimate pair correlation function by adaptive method}
\usage{
  pcf_adaptive(data.ppp, lambda, maxt, bw, adaptive = T,
    kerneltype = 1)
}
\description{
  \code{pcf_adaptive} is only a warp function of
  \link{mygest} function in InhomCluster package to fit the
  formate of fv object defined in spatstat package. See
  details of the parameters in \link{mygest}.
}

