\name{removeRareSpecies}
\alias{removeRareSpecies}
\title{Remove rare species from the community}
\usage{
  removeRareSpecies(com, maxN = 10)
}
\arguments{
  \item{com}{a scp object representing community}

  \item{maxN}{the maximum number of individuals to be a
  rare species}
}
\description{
  Remove rare species from the community
}

