\name{removeSpecies}
\alias{removeSpecies}
\title{Remvoe species from the community by a name vector}
\usage{
  removeSpecies(scp, species_names)
}
\arguments{
  \item{scp}{an scp object}

  \item{species_names}{a vector of species name that will
  be removed from the community}
}
\value{
  a scp object without the given species
}
\description{
  Remvoe species from the community by a name vector
}

