\name{rlognormal}
\alias{rlognormal}
\title{generate a lognormal species abundance distribution}
\usage{
  rlognormal(N, S, meanlog = 6, sdlog = 0.6)
}
\arguments{
  \item{N}{total number of individual}

  \item{S}{total number of species}

  \item{meanlog}{the expect mean value of lognormal
  distribution}

  \item{sdlog}{the expect standard deviation of the
  lognormal distribution}
}
\description{
  generate a lognormal species abundance distribution
}

