\name{rlogseries}
\alias{rlogseries}
\title{generate a species abundance sample from the logseries distribution given total individual N
 and total number of species S}
\usage{
  rlogseries(N, S)
}
\arguments{
  \item{N}{total number of individual}

  \item{S}{total richness}
}
\description{
  generate a species abundance sample from the logseries
  distribution given total individual N and total number of
  species S
}

