\name{SAR}
\alias{SAR}
\title{calculate the species-area relationship}
\usage{
  SAR(scp, area = 1:100, nrep = 10)
}
\arguments{
  \item{currently}{assume scp is a presence-absence
  dataframe which contains long, and lat columns. each of
  other columns are species. each row is a quadrat.}
}
\description{
  calculate the species-area relationship
}

