\name{sigHabitatTest}
\alias{sigHabitatTest}
\title{To calculate pvalues of habitat varaibles under consideration of extra aggregation existed in residual.}
\usage{
  sigHabitatTest(fittedModel, clusterResidual = TRUE)
}
\arguments{
  \item{fittedModel}{a \link{fm} object representing a
  fitted cluster point process for the given point
  pattern.}

  \item{clusterResidual}{a logical flag to consider any
  left cluster residual if it is true.}
}
\description{
  To calculate pvalues of habitat varaibles under
  consideration of extra aggregation existed in residual.
}

