\name{smallLargeDBH}
\alias{smallLargeDBH}
\title{divide the whole community into small and large DBH groups}
\usage{
  smallLargeDBH(com)
}
\arguments{
  \item{com}{a scp object}
}
\value{
  a logical vector of small individual which is true
}
\description{
  divide the whole community into small and large DBH
  groups
}

