\name{spShuffle}
\alias{spShuffle}
\title{shuffling tip label among tips of phylogeny}
\usage{
  spShuffle(phyd, n)
}
\arguments{
  \item{phyd}{a phylogenetic distance matrix with col and
  row species names}

  \item{n}{number of shuffling}
}
\value{
  a list of length n about the shuffled phylogenetic
  distance matrix
}
\description{
  shuffling tip label among tips of phylogeny
}

