\name{applyGroup}
\alias{applyGroup}
\title{apply the same function on every member of a group}
\usage{
  applyGroup(com, group, Fun, multicore = FALSE,
    mc.cores = 1, ...)
}
\arguments{
  \item{group}{factor, defination of unit of individuals
  used to fit model. e.g. species name, DBH class}
}
\description{
  apply the same function on every member of a group
}

