\name{sigHabitatTest}
\alias{sigHabitatTest}
\title{To calculate pvalues of habitat varaibles under consideration of extra aggregation existed in residual.}
\usage{
  sigHabitatTest(fittedModel)
}
\arguments{
  \item{fittedModel}{a \link{fm} object representing a
  fitted cluster point process for the given point
  pattern.}
}
\description{
  To calculate pvalues of habitat varaibles under
  consideration of extra aggregation existed in residual.
}

